v1.01; xf_rxcall improvements; various bugfixes
- Updated copyright dates
- Bumped the version to 1.01
- Added missing primitive WriteTagString()
- Updated xfiles to allow O_WRONLY as a mode.  This is forced to O_RDWR
  in modules that don't support write-only mode.
- Updated xf_rxcall to support write-only mode.
- Reworked xf_rxcall so it no longer tries to end the call; instead,
  this is always left up to the caller.  This is critical to allow
  xfiles to be used from within Rx procedure calls.  The volume-dump
  wrapper now takes care of ending its own call, so the caller API
  does not change in this case.
- Bugfix in dump - write motd and offline msg correctly
- Bugfix in parsevnode - correctly identify negative ACE's in output
- Bugfix in extract - vnode numbers are consistently afs_int32
- Bugfix: main returns int, not void
- Bugfix in xfprintf - "short" arguments are promoted to int

+++ xf_printf.c
@@ -2,7 +2,7 @@
  * CMUCS AFStools
  * dumpscan - routines for scanning and manipulating AFS volume dumps
  *
- * Copyright (c) 1998, 2001 Carnegie Mellon University
+ * Copyright (c) 1998, 2001, 2003 Carnegie Mellon University
  * All Rights Reserved.
  * 
  * Permission to use, copy, modify and distribute this software and its
@@ -264,7 +264,7 @@ afs_uint32 vxfprintf(XFILE *X, char *fmt, va_list ap)
       case 'i': 
       case 'd': /* signed decimal integer */
         if      (lflag) SVAL = va_arg(ap, long);
-        else if (hflag) SVAL = va_arg(ap, short);
+        else if (hflag) SVAL = va_arg(ap, int);
         else            SVAL = va_arg(ap, int);
         UVAL = (SVAL < 0) ? -SVAL : SVAL;
 
@@ -288,7 +288,7 @@ afs_uint32 vxfprintf(XFILE *X, char *fmt, va_list ap)
 
       case 'o': /* unsigned octal integer */
         if      (lflag) UVAL = va_arg(ap, unsigned long);
-        else if (hflag) UVAL = va_arg(ap, unsigned short);
+        else if (hflag) UVAL = va_arg(ap, unsigned int);
         else            UVAL = va_arg(ap, unsigned int);
 
         xbuf[0] = '0';
@@ -306,7 +306,7 @@ afs_uint32 vxfprintf(XFILE *X, char *fmt, va_list ap)
 
       case 'u': /* unsigned decimal integer */
         if      (lflag) UVAL = va_arg(ap, unsigned long);
-        else if (hflag) UVAL = va_arg(ap, unsigned short);
+        else if (hflag) UVAL = va_arg(ap, unsigned int);
         else            UVAL = va_arg(ap, unsigned int);
 
         if (!haveprec) {
@@ -323,7 +323,7 @@ afs_uint32 vxfprintf(XFILE *X, char *fmt, va_list ap)
       case 'x': 
       case 'X': /* unsigned hexadecimal integer */
         if      (lflag) UVAL = va_arg(ap, unsigned long);
-        else if (hflag) UVAL = va_arg(ap, unsigned short);
+        else if (hflag) UVAL = va_arg(ap, unsigned int);
         else            UVAL = va_arg(ap, unsigned int);
 
         xbuf[0] = '0';
